Modification des routes User

+++ backend/controllers/posts.js
@@ -1,8 +1,5 @@
 //Inclusion de la config
 const config = require("../config/config.json");
-const {
-    Op
-} = require("sequelize");
 
 //Import des modules
 const Models = require("../models");
@@ -11,36 +8,42 @@ const Jimp = require("jimp");
 
 //Création d'un post
 exports.createPost = (request, response) => {
-    const newPost = {
-        userId: response.locals.userId,
-        title: request.body.title,
-        image: `${request.protocol}://${request.get("host")}/images/posts/${request.file.filename}`,
-        description: request.body.description
-    }
-    //Traitement de l'image pour la limiter à 400x400 px
-    Jimp.read(`./images/posts/${request.file.filename}`).then(output => {
-        output.cover(config.jimp.postWidth, config.jimp.postWidth).write(`./images/posts/${request.file.filename}`);
-        //Création du post
-        Models.posts.create(newPost).then(() => {
-            response.status(201).json({
-                message: "Votre post a bien été publié !"
+    if (request.file) {
+        const newPost = {
+            userId: response.locals.userId,
+            title: request.body.title,
+            image: `${request.protocol}://${request.get("host")}/images/posts/${request.file.filename}`,
+            description: request.body.description
+        }
+        //Traitement de l'image pour la limiter à 400x400 px
+        Jimp.read(`./images/posts/${request.file.filename}`).then(output => {
+            output.cover(config.jimp.postWidth, config.jimp.postWidth).write(`./images/posts/${request.file.filename}`);
+            //Création du post
+            Models.Post.create(newPost).then(() => {
+                response.status(201).json({
+                    message: "Votre post a bien été publié !"
+                });
+            }).catch(error => {
+                response.status(500).json(error);
             });
         }).catch(error => {
-            response.status(500).json(error);
-        });
-    }).catch(error => {
-        FileSystem.unlink(`./images/posts/${request.file.filename}`, () => {
-            console.error(error);
-        });
+            FileSystem.unlink(`./images/posts/${request.file.filename}`, () => {
+                console.error(error);
+            });
+            response.status(500).json({
+                message: "Une erreur est survenue lors du traitement de votre image !"
+            });
+        })
+    } else {
         response.status(500).json({
-            message: "Une erreur est survenue lors du traitement de votre image !"
+            message: "Merci de renseigner une image..."
         });
-    })
+    }
 }
 
 //Récupération de tous les posts
 exports.getAllPosts = (_request, response) => {
-    Models.posts.findAll().then(result => {
+    Models.Post.findAll().then(result => {
         response.status(200).json(result);
     }).catch(error => {
         response.status(500).json(error);
@@ -49,7 +52,7 @@ exports.getAllPosts = (_request, response) => {
 
 //Récupération d'un post
 exports.getPost = (request, response) => {
-    Models.posts.findOne({
+    Models.Post.findOne({
         where: {
             id: request.params.id
         }
@@ -67,23 +70,60 @@ exports.getPost = (request, response) => {
 }
 
 //Mise à jour d'un post
-exports.updatePost = (request, response) => {}
+exports.updatePost = (request, response) => {
+    Models.Post.findOne({
+        attributes: ["id", "userId", "image", "title", "description"],
+        where: {
+            id: request.body.id
+        }
+    }).then(post => {
+        if (post) {
+            if (post.userId === response.locals.userId || response.locals.isadmin) {
+                const modifiedPost = {
+                    title: request.body.title,
+                    description: request.body.description
+                }
+                Models.Post.update(modifiedPost, {
+                    where: {
+                        id: post.id
+                    }
+                }).then(() => {
+                    response.status(200).json({
+                        message: "Post mis à jour avec succès !"
+                    });
+                }).catch(error => {
+                    response.status(500).json(error);
+                });
+            } else {
+                response.status(401).json({
+                    message: "Requête non autorisée !"
+                });
+            }
+        } else {
+            response.status(404).json({
+                message: "Ce post est introuvable !"
+            });
+        }
+    }).catch(error => {
+        response.status(500).json(error);
+    });
+}
 
 //Suppression d'un post
 exports.deletePost = (request, response) => {
-    Models.posts.findOne({
+    Models.Post.findOne({
         attributes: ["id", "userId", "image"],
         where: {
             id: request.body.id
         }
     }).then(post => {
         if (post) {
-            if (post.userId === response.locals.userId || response.locals.isadmin === true) {
+            if (post.userId === response.locals.userId || response.locals.isadmin) {
                 const img = post.image.split("/images/posts/")[1];
                 FileSystem.unlink(`images/posts/${img}`, () => {
                     console.log(`Image ${img} supprimée des resources...`);
                 });
-                Models.posts.destroy({
+                Models.Post.destroy({
                     where: {
                         id: post.id
                     }
@@ -93,16 +133,18 @@ exports.deletePost = (request, response) => {
                     });
                 }).catch(error => {
                     response.status(500).json(error);
-                })
+                });
             } else {
                 response.status(401).json({
                     message: "Requête non autorisée !"
-                })
+                });
             }
+        } else {
+            response.status(404).json({
+                message: "Ce post est introuvable !"
+            });
         }
     }).catch(error => {
-        response.status(500).json({
-            message: "Ce post est introuvable !"
-        });
+        response.status(500).json(error);
     });
 }
